refactor(httpwebrequest): merge response wrappers into one class
Fold GetResponseAsyncWrapper into GetResponseWrapper so a single class
handles GetResponse, BeginGetResponse, and EndGetResponse, mirroring
GetRequestStreamWrapper. De-duplicates TryProcessResponse and collapses
the two response tracerFactory entries into one named GetResponseWrapper
factory, retiring the vestigial legacy factory name. Behavior-preserving.
