+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/HttpWebRequest/Instrumentation.xml
@@ -20,16 +20,11 @@ SPDX-License-Identifier: Apache-2.0
 			</match>
 		</tracerFactory>
 
-		<tracerFactory name="GetResponseAsyncWrapper">
-			<match assemblyName="System" className="System.Net.HttpWebRequest">
-				<exactMethodMatcher methodName="BeginGetResponse" />
-				<exactMethodMatcher methodName="EndGetResponse" />
-			</match>
-		</tracerFactory>
-
-		<tracerFactory name="NewRelic.Agent.Core.Tracer.Factories.Net.HttpWebRequestTracerFactory">
+		<tracerFactory name="GetResponseWrapper">
 			<match assemblyName="System" className="System.Net.HttpWebRequest">
 				<exactMethodMatcher methodName="GetResponse" />
+				<exactMethodMatcher methodName="BeginGetResponse" />
+				<exactMethodMatcher methodName="EndGetResponse" />
 			</match>
 		</tracerFactory>
